% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun3.R
\name{analyze_parts_distribution}
\alias{analyze_parts_distribution}
\title{Analyzes results from the `parts_distribution' function, calculating descriptive statistics,
normality tests, fitting polynomial models, calculating probabilities
for the technological distribution, and modifying the original graph by adding a
fitting to the normal distribution for the technological data.}
\usage{
analyze_parts_distribution(parts_results, poly_degree = 3)
}
\description{
Input parameters validation according to requirements:
}
