% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Data frame of training set.}

\item{cl}{Data frame of the true class values of training data.}

\item{k_nn}{Numeric input indicating number of neighbors}

\item{k_cv}{Numeric input indicating the number of folds}
}
\value{
List object with \code{class} as a vector of predicted class
for all observations, and \code{cv_err} as a numeric with the
cross-validation misclassification error
}
\description{
This function implements the k nearest neighbors with cross validation
}
\examples{
library(dplyr)
library(class)
library(palmerpenguins)
# set random seed for reproducibility
set.seed(1234)
# omit NAs
penguins <- na.omit(penguins)
# constructing the input data frame using only the
# columns of interest
train_ <- data.frame("bill_length_mm" = penguins$bill_length_mm,
                    "bill_depth_mm" = penguins$bill_depth_mm,
                    "flipper_length_mm" = penguins$flipper_length_mm,
                    "body_mass_g" = penguins$body_mass_g)
                    # constructing the true value of class data
                    cl_ <- data.frame("species" = penguins$species)
my_knn_cv(train_, cl_, 1, 5)

}
\keyword{prediction}
