% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model Function}
\usage{
my_lm(data, formula)
}
\arguments{
\item{data}{Numeric input of the data set.}

\item{formula}{Formula input indicating the model.}
}
\value{
List including \code{model}, \code{response}, \code{coefficients},
\code{df} for degree of freedom, and \code{std_error} of the fit linear model.
}
\description{
This function fits a linear model on a given data.
}
\examples{
set.seed(1234)
# some test data
x <- rnorm(100, 0, 1)
y <- 12 * x + 5
# formula
formula <- y ~ x
data <- data.frame("x" = x, "y" = y)
# call function
my_linear_model <- my_lm(data, formula)

}
\keyword{inference}
