% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric input indicating number of folds in cross validation.}
}
\value{
Numeric with cross validation error of performing random forest,
using \code{k} folds cross validation.
}
\description{
This function implements the Random Forest function for penguins data set, with cross validation.
}
\examples{
library(randomForest)
library(dplyr)
library(class)
library(palmerpenguins)
# set random seed for reproducibility
set.seed(1234)
# omit NAs
penguins <- na.omit(penguins)
# use function
my_rf_cv(5)

}
\keyword{prediction}
