% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T hypothesis test.}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric input of the data set.}

\item{alternative}{String input indicating whether the hypothesis is "less", "greater", or "two-sided".}

\item{mu}{Numeric indicating the mean (mu) that is being tested against.}
}
\value{
List including \code{test_stat}, \code{df} for degree of freedom, \code{alternative},
and \code{p_val} of T hypothesis test.
}
\description{
This function implements one sided and two sided T hypothesis tests.
}
\examples{
set.seed(1234)
# some test data
test_data <- rnorm(100, 0, 1)
# one sided less
my_t.test(test_data, alternative = "less", mu = 0)
# one sided greater
my_t.test(test_data, alternative = "greater", mu = 0)
# two sided
my_t.test(test_data, alternative = "two.sided", mu = 0)

}
\keyword{inference}
