% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_rmst.R
\name{pseudo_rmst}
\alias{pseudo_rmst}
\title{Calculate pseudo-values for restricted mean survival time}
\usage{
pseudo_rmst(time, event, left = NULL, right = NULL, tau, censor = "right")
}
\arguments{
\item{censor}{}
}
\value{

}
\description{
Calculate pseudo-values for restricted mean survival time
}
