% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst_rc.R
\name{rmst_rc}
\alias{rmst_rc}
\title{Calculate restricted mean survival time for right censored data}
\usage{
rmst_rc(time, event, tau, subset = NULL)
}
\arguments{
\item{time}{vector of failure times}

\item{event}{vector of event status.}

\item{tau}{time horizon}

\item{subset}{vector of indices or logical vector indicating desired subset}
}
\value{
integer
}
\description{
Calculate restricted mean survival time for right censored data
}
\examples{
library(survival)
data(kidney)
s <- survfit(Surv(time, status) ~ 1, data = kidney)$surv
rmst_rc(time = kidney$time, event = kidney$status, tau = 50)
}
