% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cph.R
\name{sim_cph}
\alias{sim_cph}
\title{Simulate time-to-event Cox Proportional Hazards exponential model data}
\usage{
sim_cph(N, formula, X = NULL, beta, tfun = texp(lambda = 0.5))
}
\arguments{
\item{N}{desired sample size of simulated data}

\item{formula}{formula of covariates provided as vectors}

\item{X}{model matrix of covariates}

\item{beta}{vector of coefficients}

\item{tfun}{a time function from Bender 2005 with distributional parameters as input. Current options: \code{texp(lambda)}, \code{tweibull(lambda, nu)}, \code{tgompertz(lambda, alpha)}}
}
\value{
data.frame
}
\description{
Simulates survival data from a Cox Proportional Hazards model with the exponential distribution (Bender 2005).
}
\examples{
x1 <- rnorm(10)
x2 <- rnorm(10)
sim_cph(N = 10, formula = ~ x1 + x2, beta = c(1, 0.5),  tfun = texp(lambda = 0.5))
}
