% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texp.R
\name{texp}
\alias{texp}
\title{Generate a function that simulates time-to-event outcome from a Exponential Cox PH model}
\usage{
texp(lambda)
}
\arguments{
\item{lambda}{scale parameter (parameterization is E(X) = 1/lambda$)}
}
\value{
function that takes U and Xbeta as arguments
}
\description{
Generate a function that simulates time-to-event outcome from a Exponential Cox PH model
}
\examples{
texp(0.5)
}
