% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgompertz.R
\name{tgompertz}
\alias{tgompertz}
\title{Generate a function that simulates time-to-event outcome from a Gompertz Cox PH model}
\usage{
tgompertz(lambda, alpha)
}
\arguments{
\item{lambda}{scale parameter}

\item{alpha}{shape parameter}
}
\value{
function
}
\description{
Generate a function that simulates time-to-event outcome from a Gompertz Cox PH model
}
\examples{
tgompertz(0.5, 1)
}
