% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweibull.R
\name{tweibull}
\alias{tweibull}
\title{Generate a function that simulates time-to-event outcome from a Weibull Cox PH model}
\usage{
tweibull(lambda, nu)
}
\arguments{
\item{lambda}{scale parameter}

\item{nu}{shape parameter}
}
\value{
function
}
\description{
Generate a function that simulates time-to-event outcome from a Weibull Cox PH model
}
\examples{
tweibull(0.5, 1)
}
