% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Data frame of training set.}

\item{cl}{Data frame of the true class values of training data.}

\item{k_nn}{Numeric input indicating number of neighbors}

\item{k_cv}{Numeric input indicating the number of folds}
}
\value{
List object with \code{class} as a vector of predicted class
for all observations, and \code{cv_err} as a numeric with the
cross-validation misclassification error
}
\description{
This function implements the k nearest neighbors with cross validation
}
\examples{
my_knn_cv(train_df, class_df, 1, 5)

}
\keyword{knn}
