% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-Nearest Neighbors Cross-Validation function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{An input data frame.}

\item{cl}{True class value of your training data.}

\item{k_nn}{An integer representing the number of neighbors.}

\item{k_cv}{An integer representing the number of folds.}
}
\value{
A list with \code{class}, a vector of the predicted class Ŷi for all
observations and \code{cv_err}, a numeric with the cross-validation misclassification
error.
}
\description{
This function performs k-nearest neighbors in cross-validation.
}
\examples{
penguins_df <- my_penguins
penguins_df <- na.omit(penguins_df)
penguins_df$species <- as.character(penguins_df$species)
my_knn_cv(penguins_df, penguins_df$species, 1, 5)
}
\keyword{prediction}
