% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A \code{formula} class object, similar to \code{lm()}.}

\item{data}{An input data frame.}
}
\value{
A table with rows for each coefficient and columns for the \code{Estimate},
\code{Std. Error}, \code{t value}, and \code{Pr(>|t|)}.
}
\description{
This function fits a linear model.
}
\examples{
test <- my_gapminder
test_formula <- lifeExp ~ gdpPercap + continent
my_lm(test_formula, test)

}
\keyword{inference}
