% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Number of folds.}
}
\value{
A numeric with the cross-validation error.
}
\description{
This function performs random Forest Cross-Validation.
}
\examples{
penguins_df <- my_penguins
penguins_df <- na.omit(penguins_df)
my_rf_cv(5)
}
\keyword{prediction}
