% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{alternative}{A character string specifying the alternative hypothesis.}

\item{mu}{A number indicating the null hypothesis value of the mean.}
}
\value{
A list with elements the numeric test statistic \code{test stat},
the degrees of freedom \code{df}, the value of the parameter \code{alternative},
and the numeric p-value \code{p_val}.
}
\description{
This function performs a one sample t-test.
}
\examples{
my_t.test(1:10, "two.sided", 0)
my_t.test(1:10, "less", 1)
}
\keyword{inference}
