% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst_rc.R
\name{rmst_rc}
\alias{rmst_rc}
\title{Calculate restricted mean survival time for right censored data}
\usage{
rmst_rc(time, surv, tau)
}
\arguments{
\item{time}{vector of failure times}

\item{surv}{survival probabilities}

\item{tau}{time horizon}
}
\value{
integer
}
\description{
Calculate restricted mean survival time for right censored data
}
\examples{
#' library(survival)
data(kidney)
s <- survfit(Surv(time, status) ~ 1, data = kidney)$surv
rmst_rc(kidney$time, s, tau = 50)
}
