% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cph.R
\name{sim_cph}
\alias{sim_cph}
\title{Simulate time-to-event Cox Proportional Hazards exponential model data}
\usage{
sim_cph(N, formula, beta, lambda)
}
\arguments{
\item{N}{desired sample size of simulated data}

\item{formula}{formula of covariates}

\item{beta}{vector of coefficients}

\item{lambda}{scale parameter from the exponential distribution}
}
\value{
data.frame
}
\description{
Simulates survival data from a Cox Proportional Hazards model with the exponential distribution (Bender 2005).
}
\examples{
x1 <- rnorm(10)
x2 <- rnorm(10)
sim_cph(10, ~ x1 + x2, c(1, 0.5, 0.5), 0.50)
}
